machine-learning/llm/langchain/demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "My reference project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]

[project.scripts]
start = "main:main"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.1.0", "black>=24.0.0", "ruff>=0.1.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
